refactor(chat-tools): resolve remaining review threads
- Fix root APPPLE->APPLE typo across getSocialPlatformByLink, getPlatformDisplayName, ArtistSocial
- Standardize shared-primitive imports to named across all consumers
- Split ToolComponents into getToolCallComponent/getToolResultComponent/getToolErrorComponent (one primary export per file)
- Drop needless 'use client' from SpotifyContentCard (pure server component)
