+++ components/VercelChat/getToolResultComponent.tsx
@@ -1,4 +1,3 @@
-import { ImageSkeleton } from "@/components/VercelChat/tools/image/ImageSkeleton";
 import { ImageResult } from "@/components/VercelChat/tools/image/ImageResult";
 import {
   ImageGenerationResult,
@@ -6,7 +5,6 @@ import {
   RetrieveVideoContentResult,
 } from "@/components/VercelChat/types";
 import dynamic from "next/dynamic";
-import CreateArtistToolCall from "./tools/CreateArtistToolCall";
 import CreateArtistToolResult from "./tools/CreateArtistToolResult";
 import { CreateArtistResult } from "@/types/createArtistResult";
 import GetSpotifySearchToolResult from "./tools/GetSpotifySearchToolResult";
@@ -29,28 +27,19 @@ import { isSearchProgressUpdate } from "@/lib/search/searchProgressUtils";
 import YouTubeRevenueResult, {
   type YouTubeRevenueResult as YouTubeRevenueResultType,
 } from "./tools/youtube/YouTubeRevenueResult";
-import YouTubeRevenueSkeleton from "./tools/youtube/YouTubeRevenueSkeleton";
-import SearchWebSkeleton from "./tools/SearchWeb/SearchWebSkeleton";
-import SpotifyDeepResearchSkeleton from "./tools/SpotifyDeepResearchSkeleton";
 import { SearchApiResultType } from "./tools/SearchWeb/SearchApiResult";
 import SearchApiResult from "./tools/SearchWeb/SearchApiResult";
 import SearchWebProgress from "./tools/SearchWeb/SearchWebProgress";
 import SpotifyDeepResearchResult from "./tools/SpotifyDeepResearchResult";
 import GetArtistSocialsResult from "./tools/GetArtistSocialsResult";
-import GetArtistSocialsSkeleton from "./tools/GetArtistSocialsSkeleton";
 import GetSpotifyArtistAlbumsResult from "./tools/GetSpotifyArtistAlbumsResult";
 import { SpotifyArtistAlbumsResultUIType } from "@/types/spotify";
-import GetSpotifyArtistAlbumsSkeleton from "./tools/GetSpotifyArtistAlbumsSkeleton";
 import SpotifyArtistTopTracksResult from "./tools/SpotifyArtistTopTracksResult";
-import SpotifyArtistTopTracksSkeleton from "./tools/SpotifyArtistTopTracksSkeleton";
 import GetTasksSuccess from "./tools/tasks/GetTasksSuccess";
 import CreateTaskSuccess from "./tools/tasks/CreateTaskSuccess";
-import TasksSkeleton from "@/components/shared/TasksSkeleton";
 import GetSpotifyAlbumWithTracksResult from "./tools/GetSpotifyAlbumWithTracksResult";
-import GetSpotifyAlbumWithTracksSkeleton from "./tools/GetSpotifyAlbumWithTracksSkeleton";
 import { SpotifyAlbum } from "@/types/spotify";
 import DeleteTaskSuccess from "./tools/tasks/DeleteTaskSuccess";
-import DeleteTaskSkeleton from "./tools/tasks/DeleteTaskSkeleton";
 import UpdateTaskSuccess from "./tools/tasks/UpdateTaskSuccess";
 import { Sora2VideoSkeleton } from "./tools/sora2/Sora2VideoSkeleton";
 
@@ -61,7 +50,6 @@ const Sora2VideoResult = dynamic(
     ),
   { ssr: false, loading: () => <Sora2VideoSkeleton /> },
 );
-import CatalogSongsSkeleton from "./tools/catalog/CatalogSongsSkeleton";
 import CatalogSongsResult, {
   CatalogSongsResult as CatalogSongsResultType,
 } from "./tools/catalog/CatalogSongsResult";
@@ -71,171 +59,17 @@ import {
 } from "./tools/files/UpdateFileResult";
 import ComposioAuthResult from "./tools/composio/ComposioAuthResult";
 import { TextContent } from "@modelcontextprotocol/sdk/types.js";
-import PulseToolSkeleton from "./tools/pulse/PulseToolSkeleton";
 import PulseToolResult, {
   PulseToolResultType,
 } from "./tools/pulse/PulseToolResult";
-import GetChatsSkeleton from "./tools/chats/GetChatsSkeleton";
 import GetChatsResult, {
   GetChatsResultType,
 } from "./tools/chats/GetChatsResult";
-import RunPageSkeleton from "@/components/TasksPage/Run/RunPageSkeleton";
 import RunSandboxCommandResultWithPolling from "./tools/sandbox/RunSandboxCommandResultWithPolling";
-import ToolError from "./tools/shared/ToolError";
-import ToolStatusPill from "./tools/shared/ToolStatusPill";
 
 type CallToolResult = {
   content: TextContent[];
 };
-
-/**
- * Unified error surface for any tool that resolves to `output-error`.
- * Without this, failed tools fell through to the loading skeleton and
- * appeared to hang forever.
- */
-export function getToolErrorComponent(
-  part: ToolUIPart | DynamicToolUIPart,
-  onRetry?: () => void,
-) {
-  const { toolCallId } = part;
-  const toolName = getToolOrDynamicToolName(part);
-  // `errorText` is a runtime-populated field not yet in ToolUIPart's public type.
-  const errorText = (part as { errorText?: string }).errorText;
-  return (
-    <div key={toolCallId}>
-      <ToolError title={toolName} message={errorText} onRetry={onRetry} />
-    </div>
-  );
-}
-
-export function getToolCallComponent(part: ToolUIPart | DynamicToolUIPart) {
-  const { toolCallId } = part;
-  const toolName = getToolOrDynamicToolName(part);
-  const isSearchWebTool = toolName === "search_web";
-
-  if (toolName === "generate_image" || toolName === "edit_image") {
-    return (
-      <div key={toolCallId} className="skeleton">
-        <ImageSkeleton />
-      </div>
-    );
-  } else if (toolName === "create_new_artist") {
-    return (
-      <div key={toolCallId}>
-        <CreateArtistToolCall />
-      </div>
-    );
-  } else if (toolName === "get_youtube_revenue") {
-    return (
-      <div key={toolCallId}>
-        <YouTubeRevenueSkeleton />
-      </div>
-    );
-  } else if (isSearchWebTool) {
-    return (
-      <div key={toolCallId}>
-        <SearchWebSkeleton />
-      </div>
-    );
-  } else if (toolName === "spotify_deep_research") {
-    return (
-      <div key={toolCallId}>
-        <SpotifyDeepResearchSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_spotify_artist_albums") {
-    return (
-      <div key={toolCallId}>
-        <GetSpotifyArtistAlbumsSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_artist_socials") {
-    return (
-      <div key={toolCallId}>
-        <GetArtistSocialsSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_spotify_artist_top_tracks") {
-    return (
-      <div key={toolCallId}>
-        <SpotifyArtistTopTracksSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_tasks") {
-    return (
-      <div key={toolCallId}>
-        <TasksSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_spotify_album") {
-    return (
-      <div key={toolCallId}>
-        <GetSpotifyAlbumWithTracksSkeleton />
-      </div>
-    );
-  } else if (toolName === "create_task") {
-    return (
-      <div key={toolCallId}>
-        <TasksSkeleton />
-      </div>
-    );
-  } else if (toolName === "delete_task") {
-    return (
-      <div key={toolCallId}>
-        <DeleteTaskSkeleton />
-      </div>
-    );
-  } else if (toolName === "update_task") {
-    return (
-      <div key={toolCallId}>
-        <TasksSkeleton numberOfTasks={1} />
-      </div>
-    );
-  } else if (toolName === "retrieve_sora_2_video_content") {
-    return (
-      <div key={toolCallId}>
-        <Sora2VideoSkeleton />
-      </div>
-    );
-  } else if (
-    toolName === "insert_catalog_songs" ||
-    toolName === "select_catalog_songs"
-  ) {
-    return (
-      <div key={toolCallId}>
-        <CatalogSongsSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_pulses" || toolName === "update_pulse") {
-    return (
-      <div key={toolCallId}>
-        <PulseToolSkeleton />
-      </div>
-    );
-  } else if (toolName === "get_chats") {
-    return (
-      <div key={toolCallId}>
-        <GetChatsSkeleton />
-      </div>
-    );
-  } else if (
-    toolName === "get_task_run_status" ||
-    toolName === "prompt_sandbox"
-  ) {
-    return (
-      <div key={toolCallId}>
-        <RunPageSkeleton />
-      </div>
-    );
-  }
-
-  // Default for other tools
-  return (
-    <div key={toolCallId}>
-      <ToolStatusPill label={getToolInfo(toolName).runningLabel} />
-    </div>
-  );
-}
 
 export function getToolResultComponent(part: ToolUIPart | DynamicToolUIPart) {
   const { toolCallId, output, type } = part;
@@ -430,10 +264,3 @@ export function getToolResultComponent(part: ToolUIPart | DynamicToolUIPart) {
     />
   );
 }
-
-export const ToolComponents = {
-  getToolCallComponent,
-  getToolResultComponent,
-};
-
-export default ToolComponents;
