+++ components/VercelChat/tools/GetArtistSocialsSkeleton.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Users } from "lucide-react";
-import ToolCardSkeleton from "./shared/ToolCardSkeleton";
+import { ToolCardSkeleton } from "./shared/ToolCardSkeleton";
 
 const GetArtistSocialsSkeleton = ({ title }: { title?: string }) => {
   return (
