+++ components/VercelChat/tools/catalog/CatalogSongsSkeleton.tsx
@@ -1,5 +1,5 @@
 import { ListMusic } from "lucide-react";
-import ToolCardSkeleton from "../shared/ToolCardSkeleton";
+import { ToolCardSkeleton } from "../shared/ToolCardSkeleton";
 
 export default function CatalogSongsSkeleton() {
   return (
