+++ components/VercelChat/tools/tasks/DeleteTaskSkeleton.tsx
@@ -1,5 +1,5 @@
 import { Trash2 } from "lucide-react";
-import ToolCardSkeleton from "../shared/ToolCardSkeleton";
+import { ToolCardSkeleton } from "../shared/ToolCardSkeleton";
 
 const DeleteTaskSkeleton = () => {
   return <ToolCardSkeleton icon={Trash2} label="Deleting task" rows={1} />;
